bot2/requirements.txt
python-telegram-bot==20.7
yt-dlp==2025.01.10
requests==2.32.3
aiohttp==3.10.5
certifi==2024.8.30
urllib3==2.2.3
idna==3.10
charset-normalizer==3.3.2
